added aliases

+++ .zshrc
@@ -52,19 +52,27 @@ bindkey '^[[1;5C' forward-word                    #
 bindkey '^H' backward-kill-word                   # delete previous word with ctrl+backspace
 bindkey '^[[Z' undo                               # Shift+tab undo last action
 
-## Alias section 
-alias cp="cp -i"                                  # Confirm before overwriting something
-alias df='df -h'                                  # Human-readable sizes
-alias free='free -m'                              # Show sizes in MB
-alias gitu='git add . && git commit && git push'
-alias ll='grc ls -lha'
-alias slog='grc sudo tail -f /var/log/syslog'
-alias vim='nvim'
-#alias tmux="tmux -2"
-#alias tmux="TERM=screen-256color-bce tmux"
-#alias tmux="TERM=xterm-256color tmux"
-#alias tmux="TERM=tmux-256color tmux"
-#alias ssh='TERM=xterm-color ssh'                  # Force xterm-color on ssh sessions
+
+if [ -f ~/.config/extend-rc/aliases ]; then
+   echo "source aliases"
+    . ~/.config/extend-rc/aliases
+else
+
+   ## Alias section 
+   alias cp="cp -i"                                  # Confirm before overwriting something
+   alias df='df -h'                                  # Human-readable sizes
+   alias free='free -m'                              # Show sizes in MB
+   alias gitu='git add . && git commit && git push'
+   alias ll='grc ls -lha'
+   alias slog='grc sudo tail -f /var/log/syslog'
+   alias vim='nvim'
+   #alias tmux="tmux -2"
+   #alias tmux="TERM=screen-256color-bce tmux"
+   #alias tmux="TERM=xterm-256color tmux"
+   #alias tmux="TERM=tmux-256color tmux"
+   #alias ssh='TERM=xterm-color ssh'                  # Force xterm-color on ssh sessions
+
+fi
 
 # Theming section  
 autoload -U compinit colors zcalc
